Algunas refactorizaciones y ajustes
