+++ app/src/main/AndroidManifest.xml
@@ -11,17 +11,17 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".EditProfileActivity"></activity>
-        <activity android:name=".SplashActivity">
+        <activity android:name=".loginin.EditProfileActivity"></activity>
+        <activity android:name=".init.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".MainActivity" />
+        <activity android:name=".init.LoginActivity" />
+        <activity android:name=".init.RegisterActivity" />
+        <activity android:name=".loginin.MainActivity" />
     </application>
 
 </manifest>
