+++ app/src/main/java/com/example/projecta/init/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.projecta;
+package com.example.projecta.init;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
@@ -15,6 +15,8 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.projecta.R;
+import com.example.projecta.loginin.MainActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
