+++ app/src/main/java/com/example/projecta/init/RegisterActivity.java
@@ -1,4 +1,4 @@
-package com.example.projecta;
+package com.example.projecta.init;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,6 +14,8 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.projecta.R;
+import com.example.projecta.loginin.MainActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
