+++ app/src/main/java/com/example/projecta/loginin/fragments/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.projecta;
+package com.example.projecta.loginin.fragments;
 
 import android.os.Bundle;
 
@@ -8,13 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.projecta.R;
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class ListaFragment extends Fragment {
+public class HomeFragment extends Fragment {
 
-    public ListaFragment() {
+    public HomeFragment() {
         // Required empty public constructor
     }
 
@@ -23,6 +24,6 @@ public class ListaFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_lista, container, false);
+        return inflater.inflate(R.layout.fragment_home, container, false);
     }
 }
