+++ app/src/main/java/com/example/projecta/loginin/fragments/NavegadorFragment.java
@@ -1,4 +1,4 @@
-package com.example.projecta;
+package com.example.projecta.loginin.fragments;
 
 import android.os.Bundle;
 
@@ -10,12 +10,13 @@ import android.view.ViewGroup;
 
 import com.example.projecta.R;
 
+
 /**
  * A simple {@link Fragment} subclass.
  */
-public class HomeFragment extends Fragment {
+public class NavegadorFragment extends Fragment {
 
-    public HomeFragment() {
+    public NavegadorFragment() {
         // Required empty public constructor
     }
 
@@ -24,6 +25,6 @@ public class HomeFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false);
+        return inflater.inflate(R.layout.fragment_lista, container, false);
     }
 }
