+++ app/src/main/res/layout/activity_edit_profile.xml
@@ -5,7 +5,7 @@
     android:background="@color/colorPrimaryDark"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditProfileActivity">
+    tools:context=".loginin.EditProfileActivity">
 
     <Button
         android:id="@+id/btnCancelar"
