+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/primary_text"
-    tools:context=".LoginActivity">
+    tools:context=".init.LoginActivity">
 
     <EditText
         android:id="@+id/etPasswordLogin"
