+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:background="@color/colorPrimaryDark"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".loginin.MainActivity">
 
     <FrameLayout
         android:id="@+id/fragmentContainer"
