+++ app/src/main/res/layout/activity_splash.xml
@@ -5,6 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/splash"
-    tools:context=".SplashActivity">
+    tools:context=".init.SplashActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
